+++ src/test/java/study/stringcalculator/StringAddCalculatorTest.java
@@ -1,8 +1,7 @@
-package study.step2;
+package study.stringcalculator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import study.StringCalculator.StringAddCalculator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
